fix: the way to import contentLayout style
