+++ package.json
@@ -15,6 +15,7 @@
     "next": "10.2.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "sass": "^1.34.0",
     "swr": "^0.5.6"
   },
   "devDependencies": {
