+++ styles/contentLayout.css
@@ -1,0 +1,5 @@
+.overlay-background {
+  background-color: #fff;
+  padding: 12px;
+  border-radius: 12px;
+}
